fix: split components into files, add classname
